+++ codec/src/lib.rs
@@ -725,6 +725,9 @@ pub struct SetWindowWorkspace {
     pub workspace: String,
 }
 
+/// This is used both as a notification from server->client
+/// and as a configuration request from client->server when
+/// the client's preferred configuration changes
 #[derive(Deserialize, Serialize, PartialEq, Debug)]
 pub struct SetPalette {
     pub pane_id: PaneId,