+++ term/src/config.rs
@@ -1,4 +1,5 @@
 use crate::color::ColorPalette;
+use downcast_rs::{impl_downcast, Downcast};
 use termwiz::cell::UnicodeVersion;
 use termwiz::surface::{Line, SequenceNo};
 use wezterm_bidi::ParagraphDirectionHint;
@@ -131,7 +132,7 @@ impl Default for NewlineCanon {
 /// The configuration can be changed at runtime; provided that the implementation
 /// increments the generation counter appropriately, the changes will be detected
 /// and applied at the next appropriate opportunity.
-pub trait TerminalConfiguration: std::fmt::Debug + Send + Sync {
+pub trait TerminalConfiguration: Downcast + std::fmt::Debug + Send + Sync {
     /// Returns a generation counter for the active
     /// configuration.  If the implementation may be
     /// changed at runtime, it must increment the generation
@@ -222,6 +223,7 @@ pub trait TerminalConfiguration: std::fmt::Debug + Send + Sync {
         false
     }
 }
+impl_downcast!(TerminalConfiguration);
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct BidiMode {