+++ wezterm-client/src/client.rs
@@ -1256,4 +1256,5 @@ impl Client {
     rpc!(set_window_workspace, SetWindowWorkspace, UnitResponse);
     rpc!(set_focused_pane_id, SetFocusedPane, UnitResponse);
     rpc!(get_image_cell, GetImageCell, GetImageCellResponse);
+    rpc!(set_configured_palette_for_pane, SetPalette, UnitResponse);
 }